reduced README size, added link to goals document
If new users stumble across the repository, it seems to be useful to
direct them towards creaing user stories, which will be helpful in the
design process of the metadata structures.

The already compiled naming conventions are moved to a separate
document, which can be taken out later during the implementation phase.